Still managed to get 2 bugs. I used a bool type for error which should have been a byte. This resulted in an incorrect error message, though the flow kept working.
I also made a slip-up in the state machine. I swapped the nextStates behind state enterDigit5. So you would lose the gamer after filling in the 5th correct number

I also have changed the some of the times. It turns out 250ms to enter a number is just a bit tooo short. This was nog bug but a fine tuning